requirements.txt
absl-py==1.4.0
astor==0.8.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
colorama==0.4.6
cycler==0.11.0
fonttools==4.38.0
gast==0.5.3
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
importlib-metadata==6.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.5.5.62
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.0
pydicom==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
